examples/environmental_monitor/requirements.txt
bme680-python>=1.1.1
neo4j>=5.0.0
tensorrt-llm>=0.5.0
torch>=2.0.0
python-dotenv>=0.19.0